Using File System to Upload and Download the Files
Here i am Using fs and path to download and Upload the images from my desktop to the web instead of using the URL. This is done with Authentication check if the user is logged in then he can download that invoice file. Later i will study and implement the restriction in file accessing and generating custom pdf according to the Order details.